+++ frontend/.angular/cache/17.3.10/vite/deps/@angular_common_http.js
@@ -36,9 +36,8 @@ import {
   withNoXsrfProtection,
   withRequestsMadeViaParent,
   withXsrfConfiguration
-} from "./chunk-IQVA6OF4.js";
-import "./chunk-6SIXLUFL.js";
-import "./chunk-OBCKPMKI.js";
+} from "./chunk-A3LTE4ZH.js";
+import "./chunk-LGGNAV7W.js";
 import "./chunk-ASLTLD6L.js";
 export {
   FetchBackend,
