+++ frontend/.angular/cache/17.3.10/vite/deps/@angular_core.js
@@ -451,8 +451,7 @@ import {
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery,
   ɵɵviewQuerySignal
-} from "./chunk-6SIXLUFL.js";
-import "./chunk-OBCKPMKI.js";
+} from "./chunk-LGGNAV7W.js";
 import "./chunk-ASLTLD6L.js";
 export {
   ANIMATION_MODULE_TYPE,
