+++ frontend/.angular/cache/17.3.10/vite/deps/@angular_platform-browser.js
@@ -35,12 +35,11 @@ import {
   provideProtractorTestingSupport,
   withHttpTransferCacheOptions,
   withNoHttpTransferCache
-} from "./chunk-ZS6ZUBQH.js";
+} from "./chunk-QIQS55LB.js";
 import {
   getDOM
-} from "./chunk-IQVA6OF4.js";
-import "./chunk-6SIXLUFL.js";
-import "./chunk-OBCKPMKI.js";
+} from "./chunk-A3LTE4ZH.js";
+import "./chunk-LGGNAV7W.js";
 import "./chunk-ASLTLD6L.js";
 export {
   BrowserModule,
