+++ frontend/.angular/cache/17.3.10/vite/deps/pixi__js.js
@@ -1,37 +1,4 @@
 import {
-  BindGroupSystem,
-  GpuBatchAdaptor,
-  GpuBlendModesToPixi,
-  GpuBufferSystem,
-  GpuColorMaskSystem,
-  GpuDeviceSystem,
-  GpuEncoderSystem,
-  GpuGraphicsAdaptor,
-  GpuMeshAdapter,
-  GpuMipmapGenerator,
-  GpuRenderTarget,
-  GpuRenderTargetAdaptor,
-  GpuRenderTargetSystem,
-  GpuShaderSystem,
-  GpuStateSystem,
-  GpuStencilSystem,
-  GpuTextureSystem,
-  GpuUboSystem,
-  GpuUniformBatchPipe,
-  PipelineSystem,
-  UboBatch,
-  WGSL_ALIGN_SIZE_DATA,
-  WebGPURenderer,
-  blockDataMap,
-  createUboElementsWGSL,
-  createUboSyncFunctionWGSL,
-  generateArraySyncWGSL,
-  gpuUploadBufferImageResource,
-  gpuUploadCompressedTextureResource,
-  gpuUploadImageResource,
-  gpuUploadVideoResource
-} from "./chunk-EMKP3YEF.js";
-import {
   BUFFER_TYPE,
   GL_FORMATS,
   GL_TARGETS,
@@ -93,73 +60,6 @@ import {
   unpremultiplyAlpha,
   wrapModeToGlAddress
 } from "./chunk-3JWMVLXI.js";
-import {
-  AbstractRenderer,
-  AlphaMaskPipe,
-  ApplicationInitHook,
-  BackgroundSystem,
-  BatcherPipe,
-  BlendModePipe,
-  BufferResource,
-  CLEAR,
-  ColorMaskPipe,
-  CustomRenderPipe,
-  DATA_URI,
-  ExtractSystem,
-  Filter,
-  GenerateTextureSystem,
-  GlobalUniformSystem,
-  GpuStencilModesToPixi,
-  HelloSystem,
-  MaskFilter,
-  RenderGroupPipe,
-  RenderGroupSystem,
-  RenderTarget,
-  RenderTargetSystem,
-  RenderTexture,
-  RenderableGCSystem,
-  RendererInitHook,
-  SchedulerSystem,
-  SharedRenderPipes,
-  SharedSystems,
-  SpritePipe,
-  StencilMaskPipe,
-  SystemRunner,
-  TextureGCSystem,
-  UboSystem,
-  VERSION,
-  ViewSystem,
-  autoDetectEnvironment,
-  buildInstructions,
-  calculateProjection,
-  clearList,
-  collectAllRenderables,
-  collectRenderGroups,
-  createUboSyncFunction,
-  ensureAttributes,
-  executeInstructions,
-  fragment,
-  getCanvasTexture,
-  hasCachedCanvasTexture,
-  isRenderingToScreen,
-  loadEnvironmentExtensions,
-  mixColors,
-  mixHexColors,
-  mixStandardAnd32BitColors,
-  sayHello,
-  source,
-  textureBit,
-  textureBitGl,
-  uboSyncFunctionsSTD40,
-  uboSyncFunctionsWGSL,
-  uniformParsers,
-  unsafeEvalSupported,
-  updateRenderGroupTransform,
-  updateRenderGroupTransforms,
-  updateTransformAndChildren,
-  validateRenderables,
-  vertex
-} from "./chunk-ZT4WG52D.js";
 import {
   AccessibilitySystem,
   EventBoundary,
@@ -300,6 +200,106 @@ import {
   transformVertices,
   triangulateWithHoles
 } from "./chunk-XYGGB2JN.js";
+import {
+  BindGroupSystem,
+  GpuBatchAdaptor,
+  GpuBlendModesToPixi,
+  GpuBufferSystem,
+  GpuColorMaskSystem,
+  GpuDeviceSystem,
+  GpuEncoderSystem,
+  GpuGraphicsAdaptor,
+  GpuMeshAdapter,
+  GpuMipmapGenerator,
+  GpuRenderTarget,
+  GpuRenderTargetAdaptor,
+  GpuRenderTargetSystem,
+  GpuShaderSystem,
+  GpuStateSystem,
+  GpuStencilSystem,
+  GpuTextureSystem,
+  GpuUboSystem,
+  GpuUniformBatchPipe,
+  PipelineSystem,
+  UboBatch,
+  WGSL_ALIGN_SIZE_DATA,
+  WebGPURenderer,
+  blockDataMap,
+  createUboElementsWGSL,
+  createUboSyncFunctionWGSL,
+  generateArraySyncWGSL,
+  gpuUploadBufferImageResource,
+  gpuUploadCompressedTextureResource,
+  gpuUploadImageResource,
+  gpuUploadVideoResource
+} from "./chunk-EMKP3YEF.js";
+import {
+  AbstractRenderer,
+  AlphaMaskPipe,
+  ApplicationInitHook,
+  BackgroundSystem,
+  BatcherPipe,
+  BlendModePipe,
+  BufferResource,
+  CLEAR,
+  ColorMaskPipe,
+  CustomRenderPipe,
+  DATA_URI,
+  ExtractSystem,
+  Filter,
+  GenerateTextureSystem,
+  GlobalUniformSystem,
+  GpuStencilModesToPixi,
+  HelloSystem,
+  MaskFilter,
+  RenderGroupPipe,
+  RenderGroupSystem,
+  RenderTarget,
+  RenderTargetSystem,
+  RenderTexture,
+  RenderableGCSystem,
+  RendererInitHook,
+  SchedulerSystem,
+  SharedRenderPipes,
+  SharedSystems,
+  SpritePipe,
+  StencilMaskPipe,
+  SystemRunner,
+  TextureGCSystem,
+  UboSystem,
+  VERSION,
+  ViewSystem,
+  autoDetectEnvironment,
+  buildInstructions,
+  calculateProjection,
+  clearList,
+  collectAllRenderables,
+  collectRenderGroups,
+  createUboSyncFunction,
+  ensureAttributes,
+  executeInstructions,
+  fragment,
+  getCanvasTexture,
+  hasCachedCanvasTexture,
+  isRenderingToScreen,
+  loadEnvironmentExtensions,
+  mixColors,
+  mixHexColors,
+  mixStandardAnd32BitColors,
+  sayHello,
+  source,
+  textureBit,
+  textureBitGl,
+  uboSyncFunctionsSTD40,
+  uboSyncFunctionsWGSL,
+  uniformParsers,
+  unsafeEvalSupported,
+  updateRenderGroupTransform,
+  updateRenderGroupTransforms,
+  updateTransformAndChildren,
+  validateRenderables,
+  vertex
+} from "./chunk-ZT4WG52D.js";
 import {
   CanvasPool,
   CanvasPoolClass,
