--- frontend/.angular/cache/17.3.10/vite/deps_temp_7812983c/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
