+++ frontend/src/app/app.component.ts
@@ -39,6 +39,11 @@ interface Player {
   doswiadczenie: number
 }
 
+export interface Znajomy{
+  id: number
+  nazwa: string
+}
+
 interface Fish {
   id: number;
   nazwa: string
